% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affs.R
\name{affs}
\alias{affs}
\title{affs Function}
\usage{
affs(x, na.action = na.omit)
}
\arguments{
\item{x}{gene expression matrix where the rows represent samples, and the columns represents the features}
}
\value{
affinity graph
}
\description{
constructing  local scaling affinity
}
\examples{
data(GSE62254_data)
data1<-apply(GSE62254_gene.expression, 2, function(x) (x-mean(x))/sd(x))
affs(data1)
}
