% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNumberOfClustersGivenGraph.R
\name{estimateNumberOfClustersGivenGraph}
\alias{estimateNumberOfClustersGivenGraph}
\title{This function estimates the number of clusters using eigengap method.}
\usage{
estimateNumberOfClustersGivenGraph(W, NUMC = 2:7)
}
\arguments{
\item{W}{the similarity graph.}

\item{NUMC}{a vector which contains the possible choices of number of clusters.}
}
\value{
K1 is the estimated best number of clusters according to eigen-gaps. K12 is the estimated SECOND best number of clusters according to eigen-gaps.
}
\description{
This function estimates the number of clusters using eigengap method.
}
\examples{
data(GSE62254_data)
data1<-apply(GSE62254_gene.expression, 2, function(x) (x-mean(x))/sd(x))
A<-affs(data1)
[K1,K12]<-estimateNumberOfClustersGivenGraph(A,2:7);
}
