% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.plot.R
\name{km.plot}
\alias{km.plot}
\title{KM plot}
\usage{
km.plot(time, event, label, xlab, ylab)
}
\arguments{
\item{time}{A numeric vector representing the survival time  of a set of samples}

\item{event}{A numeric vector representing the survival status of a set of samples.}

\item{label}{the subtypes label of each sample}

\item{xlab}{}

\item{ylab}{}
}
\description{
KM plot
}
\examples{
data(GSE62254_data)
data1<-apply(GSE62254_gene.expression, 2, function(x) (x-mean(x))/sd(x))
A<-affs(data1)
Sda<-self.diffusion(A,5)
label<-spectralClustering(Sda,4)
km.plot(GSE62254_survival_time,GSE62254_survival_event,label,xlab="time",ylab = "OS")
}
