% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralClustering.R
\name{spectralClustering}
\alias{spectralClustering}
\title{spectralClustering}
\usage{
spectralClustering(affinity, K, type = 3)
}
\arguments{
\item{affinity}{input graph}

\item{K}{the number of clusters}

\item{type}{indicators of variants of spectral clustering}
}
\value{
output label
}
\description{
spectral clustering
}
\examples{
data(GSE62254_data)
data1<-apply(GSE62254_gene.expression, 2, function(x) (x-mean(x))/sd(x))
A<-affs(data1)
Sda<-self.diffusion(A,5)
spectralClustering(Sda,4)
}
\keyword{K}
\keyword{affinity}
